+++ Library/Aphid/UI/Controls.js
@@ -5,5 +5,6 @@
 Aphid.UI.Controls = {};
 
 //= require "Controls/SegmentedControl"
+//= require "Controls/ButtonControl"
 //= require "Controls/SelectView"
 //= require "Controls/DateView"